+++ src/xloft/__init__.py
@@ -3,12 +3,17 @@
 Modules exported by this package:
 
 - `namedtuple`: Class imitates the behavior of the _named tuple_.
-- `humanism` - A collection of instruments for converting data to format is convenient for humans.
+- `human` - A collection of instruments for converting data to format is convenient for humans.
 - `quantum` - A set of tools for quantum calculations.
+              Hint: uv add xloft[quantum]
 """
 
 from __future__ import annotations
 
-__all__ = ("NamedTuple",)
+__all__ = (
+    "to_human_size",
+    "NamedTuple",
+)
 
+from xloft.human import to_human_size
 from xloft.namedtuple import NamedTuple